requirements.txt
asgiref==3.7.2
channels==4.0.0
Django==4.2.10
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==3.8.0
django-simpleui==2023.3.1


#before the command run
# asgiref==3.7.2
# Django==4.2.10
# django-cors-headers==4.1.0
# django-simpleui==2023.3.1
# djangorestframework==3.14.0
# djangorestframework-simplejwt==5.2.2
# PyJWT==1.7.1
# pytz==2024.1
# sqlparse==0.4.4
# typing_extensions==4.9.0
# tzdata==2024.1